requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
cryptography
azure-functions
azure-storage-blob
azure-identity
requests
pandas
numpy
azure-ai-formrecognizer
azure-search-documents
aiohttp